+++ tests/testthat/test-relogit.R
@@ -83,3 +83,26 @@ test_that("REQUIRE TEST relogit follows ISQ (2001, eq. 11)", {
     z.out.vcov_not_robust %>% setx() %>% sim() %>% plot()
 })
 
+# REQUIRE TEST Odds Ratio summary ----------------------------------------------
+test_that('REQUIRE TEST Odds Ratio summary', {
+    data(mid)
+    z.out1 <- zelig(conflict ~ major + contig + power + maxdem + mindem + years,
+                    data = mid, model = "relogit", tau = 1042/303772,
+                    cite = FALSE, case.control = "weighting")
+
+    sum_weighting <- summary(z.out1, odds_ratios = FALSE)
+    sum_or_weighting <- summary(z.out1, odds_ratios = TRUE)
+    expect_false(sum_weighting$coefficients[1, 1] ==
+                     sum_or_weighting$coefficients[1, 1])
+    expect_equal(colnames(sum_or_weighting$coefficients)[2],
+                 "Std. Error (OR, robust)")
+
+    z.out2 <- zelig(conflict ~ major + contig + power + maxdem + mindem + years,
+                    data = mid, model = "relogit", tau = 1042/303772,
+                    cite = FALSE, case.control = "prior")
+
+    sum_weighting2 <- summary(z.out2, odds_ratios = FALSE)
+    sum_or_weighting2 <- summary(z.out2, odds_ratios = TRUE)
+    expect_equal(colnames(sum_or_weighting2$coefficients)[2],
+                 "Std. Error (OR)")
+})